feat: update claim page for legacy fees
