+++ src/app/index.tsx
@@ -11,7 +11,7 @@ import ApplicationUpdater from 'store/application/updater';
 import TransactionUpdater from 'store/transactions/updater';
 
 import { Banner } from './components/Banner';
-import { Airdrip } from './containers/Airdrip/Loadable';
+import { Claim } from './containers/Claim/Loadable';
 import Message from './Message';
 import Routes from './Routes';
 
@@ -48,7 +48,7 @@ export function App() {
           htmlAttributes={{ lang: i18n.language }}
         />
         <Switch>
-          <Route exact path="/claim" component={Airdrip} />
+          <Route exact path="/claim" component={Claim} />
           <Route component={Routes} />
         </Switch>
       </ThemeProvider>
